poprawki v2

+++ src/main/java/edu/agh/project/daos/ExaminationDao.java
@@ -55,7 +55,9 @@ public class ExaminationDao extends Dao {
     public void delete(int id) {
         beginTransaction();
         Examination examination = find(id);
-        em.remove(examination);
+        if(examination != null) {
+            em.remove(examination);
+        }
         commitTransaction();
     }
 }
